wizard: move some common code to new package "freifunk-berlin-lua-tools"
Move some generic code from the assistent into this new common package.
This way the functions can be used by other code as well.

* remove_list_entry()
* logger()
* getMergedConfig()
* mergeInto()
