+++ addons/freifunk-berlin-lua-tools/Makefile
@@ -1,0 +1,16 @@
+
+include $(TOPDIR)/rules.mk
+
+PKG_RELEASE:=1
+
+LUCI_TITLE:=Freifunk Berlin lua tools
+LUCI_DESCRIPTION:=common helper tools used by Freifunk-Berlin packages
+LUCI_DEPENDS:=luci-lib-base
+
+define Package/freifunk-berlin-lua-tools/postinst
+# noop
+endef
+
+include ../../freifunk-berlin-generic.mk
+
+# call BuildPackage - OpenWrt buildroot signature
